+++ src/main/kotlin/com/travel/travelapi/models/PlaceApi.kt
@@ -19,7 +19,7 @@ class PlaceApi(val placeId: String? = null,
                city: String? = null,
                phoneNumber: String? = null,
                website: String? = null,
-               overallStarRating: Double? = null): Place(name, description, hasSchedule, isPublic, isVerified, averageTimeSpent, latitude, longitude, address ,country, city, phoneNumber, website, overallStarRating){
+               overallStarRating: Float? = null): Place(name, description, hasSchedule, isPublic, isVerified, averageTimeSpent, latitude, longitude, address ,country, city, phoneNumber, website, overallStarRating){
 
     fun merge(place: PlaceApi){
         description = place.description ?: description
